# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Foros, Mensajes y Respuestas",
        "description": "Documentación de la API para la gestión de foros, mensajes, respuestas y autenticación.",
        "contact": {
            "email": "1813370@alu.murciaeduca.es"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://catalogovideojuegos.test/api",
            "description": "Servidor Principal de la API"
        }
    ],
    "paths": {
...
        "/mensajes/{mensajeForo}/respuestas": {
            "get": {
                "tags": [
                    "Respuestas de Mensaje"
                ],
                "summary": "Obtener respuestas de un mensaje específico",
                "operationId": "getRespuestasDeMensaje",
                "parameters": [
                    {
                        "name": "mensajeForo",
                        "in": "path",
                        "description": "ID del mensaje padre",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Paginado",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/RespuestaForo"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Mensaje no encontrado"
                    }
                }
            },
            "post": {
                "tags": [
                    "Respuestas de Mensaje"
                ],
                "summary": "Crear una nueva respuesta a un mensaje",
                "operationId": "storeRespuestaForo",
                "parameters": [
                    {
                        "name": "mensajeForo",
                        "in": "path",
                        "description": "ID del mensaje al que responder",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Contenido de la respuesta",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RespuestaForoStoreRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Respuesta creada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/RespuestaForo"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "403": {
                        "description": "No autorizado (si aplica policy)"
                    },
                    "404": {
                        "description": "Mensaje padre no encontrado"
                    },
                    "422": {
                        "description": "Error de validación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorValidation"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/respuestas/{respuestaForo}": {
            "get": {
                "tags": [
                    "Respuestas de Mensaje"
                ],
                "summary": "Obtener una respuesta específica",
                "operationId": "getRespuestaForoById",
                "parameters": [
                    {
                        "name": "respuestaForo",
                        "in": "path",
                        "description": "ID de la respuesta",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operación exitosa",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/RespuestaForo"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Respuesta no encontrada"
                    }
                }
            },
            "put": {
                "tags": [
                    "Respuestas de Mensaje"
                ],
                "summary": "Actualizar una respuesta existente",
                "operationId": "updateRespuestaForo",
                "parameters": [
                    {
                        "name": "respuestaForo",
                        "in": "path",
                        "description": "ID de la respuesta a actualizar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Datos de la respuesta a actualizar. 'contenido' es el campo principal.",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RespuestaForoStoreRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Respuesta actualizada exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/RespuestaForo"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "403": {
                        "description": "No autorizado"
                    },
                    "404": {
                        "description": "Respuesta no encontrada"
                    },
                    "422": {
                        "description": "Error de validación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorValidation"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Respuestas de Mensaje"
                ],
                "summary": "Eliminar una respuesta",
                "operationId": "deleteRespuestaForo",
                "parameters": [
                    {
                        "name": "respuestaForo",
                        "in": "path",
                        "description": "ID de la respuesta a eliminar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Respuesta eliminada exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Respuesta eliminada correctamente."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "403": {
                        "description": "No autorizado"
                    },
                    "404": {
                        "description": "Respuesta no encontrada"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "UserLite": {
                "title": "User Lite Information",
                "description": "Información básica del usuario",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Nombre Usuario"
                    }
                },
                "type": "object"
            },
...
            "RespuestaForo": {
                "title": "Respuesta de Mensaje de Foro",
                "description": "Representación de una respuesta a un mensaje de foro",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 101
                    },
                    "contenido": {
                        "type": "string",
                        "example": "Esta es una respuesta."
                    },
                    "usuario": {
                        "$ref": "#/components/schemas/UserLite"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-06-04T10:30:00.000000Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-06-04T10:35:00.000000Z"
                    }
                },
                "type": "object"
            },
...
            "ErrorValidation": {
                "title": "Error de Validación",
                "description": "Objeto de error devuelto en caso de fallos de validación (422)",
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "The given data was invalid."
                    },
                    "errors": {
                        "description": "Objeto con los errores de validación por campo",
                        "type": "object",
                        "example": {
                            "titulo": [
                                "El campo título es obligatorio."
                            ]
                        }
                    }
                },
                "type": "object"
            },
...
            "RespuestaForoStoreRequest": {
                "title": "Respuesta Foro Store Request Body",
                "description": "Cuerpo de la petición para crear una respuesta de foro",
                "required": [
                    "contenido"
                ],
                "properties": {
                    "contenido": {
                        "type": "string",
                        "example": "Estoy de acuerdo con este mensaje."
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Autenticación por Token Bearer (Sanctum)",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}